Add new function for more accurate fuel use
Due to the tyranny of the rocket equation, the fuel to lift a
module requires additional fuel to lift it as well.  The new
module-fuel-required function starts with the fuel needed to
lift the module mass, then iterates on the fuel required to
lift each new block of fuel for lifting the fuel, until the
fuel requirement is zero, then adds them all up to get the
full fuel requirement for the module's mass.