Fixes for Intcode computer
Typo in load-program

Misunderstood logic for opcode 9, the value is added to the current
base, not replaces it

Mode 2 is valid for target addresses, but also works a little
differently than other values.  Regular values would use the
relative address to point to the place to read for the value.
For targets, the relative address is the "value".  Added a
get-target function and changed all target lookups on opcodes to
use it.  It treats mode 0 and 1 as the same, and does relative
address for mode 2.