Solution for Day 14 Part 1
We use a map, keyed by ingredient, that tracks the amount produced
so far and the amount requested.  Starting with 0 FUEL produced and
1 requested, we reduce over each ingredient in our list, and if
the requested amount is below the produced amount, we add the
ingredients required to product one recipe run to our requested
numbers.  We continue this process (omitting ORE, of course) until
the ingredient list stabilizing, which means all requests have been
satisfied.  From there, the number of ORE requested is our answer.