Solution for Day 14 Part 2
It turned out that adding ingredients one recipe batch at a time
was way too slow, so modified update-ingredient to calculate how
many recipe batches were needed and add them all at once.

Then, I start calculating the ORE usage for 1000 FUEL, and if it is
less than the 1 trillion threshold, add another 1000 FUEL and
recalculate.  If we pass the 1 trillion, then I divide the increment
by 10, and try again from the previous iteration's list.  When we pass
1 trillion with the increment at 1, then we know we have produced
the maximum amount of FUEL for 1 trillion ORE and return that value.