Add solution for Day 15 Part 2
For this, we need to continue past the generator, and we also want
to know the distance from the generator rather than origin.  To do
this, we use a similar driver, but when we find the generator, we
flip all the distances by subtracting them from the generator
distance.  Note that this only works because it's a single path,
with no branches.  We then continue following the left wall until
we return to origin.  At that point, the maximum distance in the
distances map is our answer.

In hindsight, I just realized that I should have initialized the
distance map when I reached the generator, but oh well.