Add solution for Day 16 Part 2
Admittedly, I did not see the trick, and went seeking tips.  Even
though I was looking for "make the math faster" tips, I ended up
finding the trick on reddit.  The key for part 2 is the offset.
Since the offset is more than halfway through the input, even
after the 10000 repetitions, all the digits before the offset
are multiplied by 0, and thus irrelevant.  At the same time,
the rest of the digits will be multiplied by 1, so the other 0
and the -1 vanish as well.  At this point, each digit is just the
sum of all the digits after it.  It probably wouldn't have been
too terrible to do the calculations straight up, but there was a
second trick of working backwards and doing partial sums along
the way.  Clojure happens to have a (reductions) function that
practically does the work for you.

Moral of the story: more thinking; less typing.