Migrate previous solutions away from async
Went back through the previous solutions that were using the async
intcode computer and updated them to work with the new version.
Luckily, it was mostly straightforward, once I determined what
each problem was supposed to be doing.

Of course, the fact that I had to go back and read the problem
descriptions sometimes to remember what the code was supposed to
do is a bit of an indictment of my self-documenting code skills.
But, as the old proverb goes, "There are only two hard problems in
Computer Science: naming things, cache invalidation, and off-by-one
errors".  The first one is demonstrated here.